requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
alsa-lib=1.2.3=h516909a_0
argon2-cffi=20.1.0=py39h3811e60_2
async_generator=1.10=py_0
attrs=21.2.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bleach=3.3.0=pyh44b312d_0
blinker=1.4=py_1
brotlipy=0.7.0=py39h3811e60_1001
ca-certificates=2021.5.30=ha878542_0
certifi=2021.5.30=py39hf3d152e_0
cffi=1.14.5=py39he32792d_0
chardet=4.0.0=py39hf3d152e_1
cryptography=3.4.7=py39hbca0aa6_0
cudatoolkit=10.2.89=h8f6ccaa_8
cycler=0.10.0=py_2
dbus=1.13.6=h48d8840_2
debugpy=1.3.0=py39he80948d_0
decorator=5.0.9=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
expat=2.4.1=h9c3ff4c_0
fontconfig=2.13.1=hba837de_1005
freetype=2.10.4=h0708190_1
gettext=0.19.8.1=h0b5b191_1005
glib=2.68.3=h9c3ff4c_0
glib-tools=2.68.3=h9c3ff4c_0
gst-plugins-base=1.18.4=hf529b03_2
gstreamer=1.18.4=h76c114f_2
icu=68.1=h58526e2_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=4.6.0=py39hf3d152e_0
intel-openmp=2021.2.0=h06a4308_610
ipykernel=6.0.0=py39hef51801_0
ipython=7.25.0=py39hef51801_1
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.3=pyhd3deb0d_0
jbig=2.1=h7f98852_2003
jedi=0.18.0=py39hf3d152e_2
jinja2=3.0.1=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter=1.0.0=py39hf3d152e_6
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_console=6.4.0=pyhd8ed1ab_0
jupyter_core=4.7.1=py39hf3d152e_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_widgets=1.0.0=pyhd8ed1ab_1
kiwisolver=1.3.1=py39h1a9c180_1
krb5=1.19.1=hcc1bbae_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_0
lerc=2.2.1=h9c3ff4c_0
libblas=3.9.0=9_mkl
libcblas=3.9.0=9_mkl
libclang=11.1.0=default_ha53f305_1
libdeflate=1.7=h7f98852_5
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.10=hcdb4288_3
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h2828fa1_19
libgfortran-ng=9.3.0=hff62375_19
libgfortran5=9.3.0=hff62375_19
libglib=2.68.3=h3e27bee_0
libgomp=9.3.0=h2828fa1_19
libiconv=1.16=h516909a_0
liblapack=3.9.0=9_mkl
libllvm11=11.1.0=hf817b99_2
libogg=1.3.4=h7f98852_1
libopenblas=0.3.15=pthreads_h8fe5266_1
libopus=1.3.1=h7f98852_1
libpng=1.6.37=h21135ba_2
libpq=13.3=hd57d9b9_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.3.0=h6de172a_19
libtiff=4.3.0=hf544144_1
libuuid=2.32.1=h7f98852_1000
libuv=1.41.0=h7f98852_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.2.0=h7f98852_2
libxcb=1.13=h7f98852_1003
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
lz4-c=1.9.3=h9c3ff4c_0
markupsafe=2.0.1=py39h3811e60_0
matplotlib=3.4.2=py39hf3d152e_0
matplotlib-base=3.4.2=py39h2fa2bec_0
matplotlib-inline=0.1.2=pyhd8ed1ab_2
mistune=0.8.4=py39h3811e60_1004
mkl=2021.2.0=h06a4308_296
mysql-common=8.0.25=ha770c72_2
mysql-libs=8.0.25=hfa10184_2
nbclient=0.5.3=pyhd8ed1ab_0
nbconvert=6.1.0=py39hf3d152e_0
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.2=h58526e2_4
nest-asyncio=1.5.1=pyhd8ed1ab_0
ninja=1.10.2=h4bd325d_0
notebook=6.4.0=pyha770c72_0
nspr=4.30=h9c3ff4c_0
nss=3.67=hb5efdd6_0
numpy=1.21.0=py39hdbf815f_0
oauthlib=3.1.1=pyhd8ed1ab_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1k=h7f98852_0
packaging=20.9=pyh44b312d_0
pandas=1.2.5=py39hde0f152_0
pandoc=2.14.0.3=h7f98852_0
pandocfilters=1.4.2=py_1
parso=0.8.2=pyhd8ed1ab_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.3.0=py39ha612740_0
pip=21.1.3=pyhd8ed1ab_0
prometheus_client=0.11.0=pyhd8ed1ab_0
prompt-toolkit=3.0.19=pyha770c72_0
prompt_toolkit=3.0.19=hd8ed1ab_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.9.0=pyhd8ed1ab_0
pyjwt=2.1.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py39hf3d152e_7
pyqt-impl=5.12.3=py39h0fcd23e_7
pyqt5-sip=4.19.18=py39he80948d_7
pyqtchart=5.12=py39h0fcd23e_7
pyqtwebengine=5.12.1=py39h0fcd23e_7
pyrsistent=0.17.3=py39h3811e60_2
pysocks=1.7.1=py39hf3d152e_3
python=3.9.5=h49503c6_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.9=2_cp39
pytorch=1.9.0=py3.9_cuda10.2_cudnn7.6.5_0
pytz=2021.1=pyhd8ed1ab_0
pyzmq=22.1.0=py39h37b5a0c_0
qt=5.12.9=hda022c4_4
qtconsole=5.1.1=pyhd8ed1ab_0
qtpy=1.9.0=py_0
readline=8.1=h46c0cb4_0
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
send2trash=1.7.1=pyhd8ed1ab_0
setuptools=49.6.0=py39hf3d152e_3
six=1.16.0=pyh6c4a22f_0
sqlite=3.36.0=h9cd32fc_0
terminado=0.10.1=py39hf3d152e_0
testpath=0.5.0=pyhd8ed1ab_0
tk=8.6.10=h21135ba_1
torchaudio=0.9.0=py39
torchvision=0.2.2=py_3
tornado=6.1=py39h3811e60_1
traitlets=5.0.5=py_0
tweepy=3.10.0=pyhd8ed1ab_0
typing_extensions=3.10.0.0=pyha770c72_0
tzdata=2021a=he74cb21_0
urllib3=1.26.6=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.36.2=pyhd3deb0d_0
widgetsnbextension=3.5.1=py39hf3d152e_4
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h516909a_1
zeromq=4.3.4=h9c3ff4c_0
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h516909a_1010
zstd=1.5.0=ha95c52a_0
